get rid of all codepaths where pp = 0, rearrange layout
