+++ megatron/model/__init__.py
@@ -16,7 +16,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .gpt2_model import GPT2Model, GPT2ModelPipe
+from .gpt2_model import GPT2ModelPipe
 from .utils import get_params_for_weight_decay_optimization
-from .language_model import get_language_model
-from .norms import RMSNorm, ScaleNorm, LayerNorm
