+++ megatron/training.py
@@ -296,6 +296,8 @@ def train_step(forward_step_func, data_iterator,
     if args.pipe_parallel_size > 0:
         return train_step_pipe(model, data_iterator)
 
+    # TODO: Dead code (?)
+
     # Forward model for one step.
     timers('forward').start()
     loss, loss_reduced = forward_step_func(data_iterator, model)
@@ -371,7 +373,9 @@ def training_log(loss_dict, total_loss_dict, learning_rate, iteration,
         if name in timers.timers:
             timers_to_log.append(name)
 
-    if args.pipe_parallel_size <= 0:
+    if args.pipe_parallel_size == 0:
+        # TODO: Dead code?
+
         add_to_logging('forward')
         add_to_logging('backward')
         add_to_logging('backward-backward')
@@ -554,6 +558,7 @@ def train(forward_step_func, model, optimizer, lr_scheduler,
             sys.exit()
 
     return iteration
+
 
 
 def evaluate(forward_step_func, data_iterator, model, verbose=False):
