+++ text_gen_gpt2.py
@@ -53,7 +53,7 @@ def main(extra_args_provider=None, get_key_value=True):
 
     # Set up model and load checkpoint.
     model, _, _ = setup_model_and_optimizer(lambda: model_provider(use_wandb=False, inference=True, get_key_value=get_key_value))
-    if args.pipe_parallel_size == 1 and isinstance(model, PipelineEngine):
+    if args.pipe_parallel_size <= 1 and isinstance(model, PipelineEngine):
         # if it's a pipe parallel model but not actually doing parallelism, convert it to a normal deepspeed model
         model = pipe_to_normal(model)
     print_rank_0('Finished loading model')
